requirements.txt
streamlit==1.29.0
openai==1.12.0
langchain==0.1.5
langchain-openai==0.0.5
langchain-community==0.0.16
faiss-cpu==1.7.4
pandas==2.1.4
numpy==1.26.3
python-dotenv==1.0.0